Added casts to fix compiler warnings.
git-svn-id: https://svn.aros.org/svn/aros/trunk@20847 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/graphics/allocbitmap.c
@@ -381,7 +381,7 @@
 
 	if (nbm)
 	{
-	    int plane;
+	    ULONG plane;
 
 	    nbm->BytesPerRow = ((sizex + 15) >> 4) * 2;
 	    nbm->Rows	     = sizey;
