+++ rom/graphics/scrollraster.c
@@ -106,9 +106,9 @@
 	struct BitMap * bm = rp->BitMap;
 	ULONG width  = GetBitMapAttr(bm, BMA_WIDTH);
 	ULONG height = GetBitMapAttr(bm, BMA_HEIGHT);
-	if (xMax >= width )
+	if ((ULONG)xMax >= width )
 	    xMax = width - 1;
-	if (yMax >= height)
+	if ((ULONG)yMax >= height)
 	    yMax = height - 1;
     }
